reload filepdf

+++ Services/Repositories/Implimentations/QuanLyHoaDon/HoaDonDienTuService.cs
@@ -35,7 +35,6 @@ using Services.ViewModels.TienIch;
 using Services.ViewModels.XML.QuyDinhKyThuatHDDT.Enums;
 using Spire.Doc;
 using Spire.Doc.Documents;
-using Spire.Doc.Fields;
 using Spire.Pdf;
 using Spire.Pdf.Graphics;
 using System;
@@ -1858,7 +1857,7 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                 string pdfFileName = string.Empty;
                 string xmlFileName = string.Empty;
 
-                if (hd.IsCapMa != true && ((hd.TrangThaiQuyTrinh == (int)TrangThaiQuyTrinh.DaKyDienTu) || (hd.TrangThaiQuyTrinh == (int)TrangThaiQuyTrinh.CQTDaCapMa)) && (!string.IsNullOrEmpty(hd.FileDaKy) || !string.IsNullOrEmpty(hd.XMLDaKy)))
+                if (hd.IsCapMa != true && hd.IsReloadSignedPDF != true && ((hd.TrangThaiQuyTrinh == (int)TrangThaiQuyTrinh.DaKyDienTu) || (hd.TrangThaiQuyTrinh == (int)TrangThaiQuyTrinh.CQTDaCapMa)) && (!string.IsNullOrEmpty(hd.FileDaKy) || !string.IsNullOrEmpty(hd.XMLDaKy)))
                 {
                     return new KetQuaConvertPDF
                     {
@@ -1897,9 +1896,9 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
 
                 doc.Replace(LoaiChiTietTuyChonNoiDung.MaTraCuu.GenerateKeyTag(), hd.MaTraCuu ?? string.Empty, true, true);
 
-                if (hd.IsCapMa == true || hd.IsPhatHanh == true)
+                if (hd.IsCapMa == true || hd.IsPhatHanh == true || hd.IsReloadSignedPDF == true)
                 {
-                    ImageHelper.AddSignatureImageToDoc(doc, hoSoHDDT.TenDonVi);
+                    ImageHelper.AddSignatureImageToDoc(doc, hoSoHDDT.TenDonVi, hd.NgayHoaDon);
                 }
                 else
                 {
@@ -2024,7 +2023,7 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
 
                 string fullPdfFolder;
                 string fullXmlFolder;
-                if (hd.IsCapMa == true)
+                if (hd.IsCapMa == true || hd.IsReloadSignedPDF == true)
                 {
                     fullPdfFolder = Path.Combine(_hostingEnvironment.WebRootPath, $"FilesUpload/{databaseName}/{ManageFolderPath.PDF_SIGNED}");
                     fullXmlFolder = Path.Combine(_hostingEnvironment.WebRootPath, $"FilesUpload/{databaseName}/{ManageFolderPath.XML_SIGNED}");
@@ -2045,18 +2044,21 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                         }
                     }
 
-                    if (!Directory.Exists(fullXmlFolder))
+                    if (hd.IsCapMa == true)
                     {
-                        Directory.CreateDirectory(fullXmlFolder);
-                    }
-                    else
-                    {
-                        if (!string.IsNullOrEmpty(hd.XMLDaKy))
+                        if (!Directory.Exists(fullXmlFolder))
                         {
-                            string oldFilePath = Path.Combine(fullXmlFolder, hd.XMLDaKy);
-                            if (File.Exists(oldFilePath))
+                            Directory.CreateDirectory(fullXmlFolder);
+                        }
+                        else
+                        {
+                            if (!string.IsNullOrEmpty(hd.XMLDaKy))
                             {
-                                File.Delete(oldFilePath);
+                                string oldFilePath = Path.Combine(fullXmlFolder, hd.XMLDaKy);
+                                if (File.Exists(oldFilePath))
+                                {
+                                    File.Delete(oldFilePath);
+                                }
                             }
                         }
                     }
@@ -2103,12 +2105,21 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
 
                 var entity = await _db.HoaDonDienTus.FirstOrDefaultAsync(x => x.HoaDonDienTuId == hd.HoaDonDienTuId);
 
-                if (hd.IsCapMa == true)
+                if (hd.IsCapMa == true || hd.IsReloadSignedPDF == true)
                 {
                     pdfFileName = $"{hd.BoKyHieuHoaDon.KyHieu}-{hd.SoHoaDon}-{Guid.NewGuid()}.pdf";
-                    xmlFileName = $"{hd.BoKyHieuHoaDon.KyHieu}-{hd.SoHoaDon}-{Guid.NewGuid()}.xml";
                     entity.FileDaKy = pdfFileName;
-                    entity.XMLDaKy = xmlFileName;
+
+                    if (hd.IsCapMa == true)
+                    {
+                        xmlFileName = $"{hd.BoKyHieuHoaDon.KyHieu}-{hd.SoHoaDon}-{Guid.NewGuid()}.xml";
+                        entity.XMLDaKy = xmlFileName;
+                    }
+                    else
+                    {
+                        xmlFileName = entity.XMLDaKy;
+                    }
+
                     entity.TrangThaiQuyTrinh = (int)TrangThaiQuyTrinh.CQTDaCapMa;
                     entity.MaCuaCQT = hd.MaCuaCQT;
                 }
@@ -2127,16 +2138,19 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
                 hd.SoTienBangChu = soTienBangChu;
                 doc.SaveToFile(fullPdfFilePath, Spire.Doc.FileFormat.PDF);
 
-                if (hd.IsCapMa == true)
+                if (hd.IsCapMa == true || hd.IsReloadSignedPDF == true)
                 {
-                    File.WriteAllText(fullXmlFilePath, hd.DataXML);
+                    if (hd.IsCapMa == true)
+                    {
+                        File.WriteAllText(fullXmlFilePath, hd.DataXML);
+                    }
                 }
                 else
                 {
                     await _xMLInvoiceService.CreateXMLInvoice(fullXmlFilePath, hd);
                 }
 
-                if (hd.IsCapMa == true)
+                if (hd.IsCapMa == true || hd.IsReloadSignedPDF == true)
                 {
                     await UpdateFileDataForHDDT(hd.HoaDonDienTuId, fullPdfFilePath, fullXmlFilePath);
                 }
@@ -2158,7 +2172,6 @@ namespace Services.Repositories.Implimentations.QuanLyHoaDon
             }
             catch (Exception e)
             {
-
                 throw;
             }
         }
